+++ test/GitHub.UI.UnitTests/GitHub.UI.UnitTests.csproj
@@ -48,6 +48,8 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AppDomainContext.cs" />
+    <Compile Include="ResourceDictionaryUtilities.cs" />
     <Compile Include="SharedDictionaryManagerTests.cs" />
     <Compile Include="LoadingResourceDictionaryTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
